% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_processing.R
\name{aoi_align}
\alias{aoi_align}
\title{Align AOI to product}
\usage{
aoi_align(aoi = NULL, product = NULL, snap = "out")
}
\arguments{
\item{aoi}{`sf` bbox object}

\item{product}{Character, one of the options from column 'Short_Name' in
\code{\link[slga:slga_product_info]{slga_product_info}}.}

\item{snap}{Character; 'near', 'in', or 'out'. Defaults to 'out'.}
}
\description{
Cheerfully stolen from code{\link[raster:alignExtent]{raster::alignExtent}}
and adapted to sf bbox objects. This is designed to prevent WCS server-side
interpolation when requesting SLGA data.
}
\keyword{internal}
