% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_processing.R
\name{aoi_convert}
\alias{aoi_convert}
\alias{aoi_convert.numeric}
\alias{aoi_convert.Raster}
\alias{aoi_convert.Extent}
\alias{aoi_convert.sf}
\alias{aoi_convert.sfc}
\alias{aoi_convert.sfg}
\alias{aoi_convert.s2_geography}
\title{Convert AOI}
\usage{
aoi_convert(aoi = NULL)

\method{aoi_convert}{numeric}(aoi = NULL)

\method{aoi_convert}{Raster}(aoi = NULL)

\method{aoi_convert}{Extent}(aoi = NULL)

\method{aoi_convert}{sf}(aoi = NULL)

\method{aoi_convert}{sfc}(aoi = NULL)

\method{aoi_convert}{sfg}(aoi = NULL)

\method{aoi_convert}{s2_geography}(aoi = NULL)
}
\arguments{
\item{aoi}{Numeric vector of bounding coordinates in xmin, ymin, xmax, ymax
order; or an `sf` or `raster` object from which they can be derived.}
}
\value{
`sf` bbox object with same crs as input.
}
\description{
Converts an AOI from a variety of possible input types to an `sf` style bbox.
}
\keyword{internal}
