% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_processing.R
\name{aoi_overlaps}
\alias{aoi_overlaps}
\title{Check AOI overlaps}
\usage{
aoi_overlaps(aoi = NULL, product = NULL)
}
\arguments{
\item{aoi}{Converted and transformed area of interest}

\item{product}{Character, one of the options from column 'Short_Name' in
\code{\link[slga:slga_product_info]{slga_product_info}}.}
}
\value{
Logical, TRUE if overlap exists.
}
\description{
Confirms whether an AOI covers the requested product (wholly or partly).
}
\note{
Will return TRUE if bbox is over product but not over land.
}
