% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_processing.R
\name{aoi_tile}
\alias{aoi_tile}
\title{Tile AOI}
\usage{
aoi_tile(aoi = NULL, product = NULL, size = 1)
}
\arguments{
\item{aoi}{sf bbox already converted from user input and aligned to requested
product.}

\item{size}{Number, max side length of tiles in decimal degrees. Default 1.}
}
\value{
a list of bounding boxes subdividing the area of interest. Return
list is aligned to product and tiles that don't overlap are discarded.
}
\description{
If an AOI is Large, chop it up so that a series of GETs can be sent.
}
\keyword{internal}
