% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_processing.R
\name{aoi_transform}
\alias{aoi_transform}
\title{Transform an AOI's CRS}
\usage{
aoi_transform(aoi = NULL, crs = 4283)
}
\arguments{
\item{aoi}{Object of class 'bbox' generated by internal function
\code{\link[slga:aoi_convert]{aoi_convert()}}.}

\item{crs}{EPSG code of desired coordinate reference system. Default 4283.}
}
\value{
'bbox' object with the desired CRS, large enough to ensure returned
dataset covers entire input object.
}
\description{
Transforms an `sf` style bounding box object into the specified coordinate
reference system.
}
\note{
This is a stopgap until and unless a dedicated sf method for bbox
transformations is written. Currently used to coerce to WGS84 under specific
circumstances. Since this helper was written for an Australia-specific package,
it is assumed that it will be used to transform coordinates from a fairly
limited list of known crs's, chiefly GDA94 (EPSG:4283) and its associated
UTM and Albers equal area projections.
}
\keyword{internal}
