% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_aoi}
\alias{convert_aoi}
\alias{convert_aoi.numeric}
\alias{convert_aoi.Raster}
\alias{convert_aoi.Extent}
\alias{convert_aoi.sf}
\alias{convert_aoi.sfc}
\title{Convert AOI}
\usage{
convert_aoi(aoi = NULL)

\method{convert_aoi}{numeric}(aoi = NULL)

\method{convert_aoi}{Raster}(aoi = NULL)

\method{convert_aoi}{Extent}(aoi = NULL)

\method{convert_aoi}{sf}(aoi = NULL)

\method{convert_aoi}{sfc}(aoi = NULL)
}
\arguments{
\item{aoi}{Numeric; bounding coordinates or an `sf` or `raster` object from
which they can be derived.}
}
\value{
`sf` bbox object with same crs as input.
}
\description{
Converts an AOI from a variety of possible input types to an `sf` style bbox.
}
\keyword{internal}
