% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_generate.R
\name{make_soils_url}
\alias{make_soils_url}
\title{Make SLGA WCS URL}
\usage{
make_soils_url(
  product = NULL,
  attribute = NULL,
  component = NULL,
  depth = NULL,
  aoi = NULL,
  req_type = "cov"
)
}
\arguments{
\item{product}{Character, one of the options from column 'Short_Name' in
\code{\link[slga:slga_product_info]{slga_product_info}}.}

\item{attribute}{Character, one of the options from column 'Code' in
\code{\link[slga:slga_attribute_info]{slga_attribute_info}}, where 'Type' =
'Soil'.}

\item{component}{Character, one of 'VAL', 'CLO', or 'CHI'.}

\item{depth}{Integer, a number from 1 to 6.}

\item{aoi}{Vector of WGS84 coordinates defining a rectangular area of
interest. The vector may be specified directly in the order xmin, ymin,
xmax, ymax, or the function can derive an aoi from the boundary of an `sf`
or `raster` object.}

\item{req_type}{Character; one of 'cap', 'cov' or 'desc'. Defaults to 'cov'.}
}
\description{
Generate the URL for a particular soils product, attribute, component and
depth available from the Soil and Landscape Grid of Australia.
}
\keyword{internal}
