% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slga.R
\docType{package}
\name{slga-package}
\alias{slga}
\alias{slga-package}
\title{slga: Soil and Landscape Grid of Australia services}
\description{
Provides access to Soil and Landscape Grid raster datasets via existing 
WCS services. See www.csiro.au/soil-and-landscape-grid.
}
\author{
\strong{Maintainer}: Lauren O'Brien \email{obrlsoilau@gmail.com}

Authors:
\itemize{
  \item Ross Searle \email{ross.searle@csiro.au}
}

}
