% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{slga_filenamer}
\alias{slga_filenamer}
\title{Filename generator}
\usage{
slga_filenamer(
  product = NULL,
  attribute = NULL,
  component = NULL,
  depth = NULL
)
}
\arguments{
\item{product}{Character, one of the options from column 'Code' in
code{\link[slga:slga_product_info]{slga_product_info}} where Type = 'Soil'.}

\item{attribute}{Character, one of the options from column 'Code' in
code{\link[slga:slga_attribute_info]{slga_attribute_info}}.}

\item{component}{Character, one of 'ALL', 'VAL', 'CIS', 'CLO', or
'CHI'. Defaults to 'ALL'.}

\item{depth}{Integer from 1 to 6.}
}
\value{
filename string
}
\description{
generates a filename for an SLGA raster request
}
\keyword{internal}
