% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_processing.R
\name{tidy_lscape_data}
\alias{tidy_lscape_data}
\title{tidy landscape parameter rasters}
\usage{
tidy_lscape_data(r = NULL, product = NULL, write_out = NULL)
}
\arguments{
\item{r}{raster object downloaded over WCS}

\item{product}{character; the SLGA terrain product in question}

\item{write_out}{Boolean, whether to write the processed dataset to the
working directory as a GeoTiff.}
}
\value{
a raster, but a better one
}
\description{
Does some post-processing on landscape datasets downloaded from SLGA - mostly
about setting NA correctly
}
\keyword{internal}
