% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_processing.R
\name{tidy_soils_data}
\alias{tidy_soils_data}
\title{Tidy soils rasters}
\usage{
tidy_soils_data(r = NULL, out_name = NULL)
}
\arguments{
\item{r}{raster object downloaded over WCS}

\item{product}{character; the SLGA soils product in question}
}
\value{
a raster, but a better one
}
\description{
Does some post-processing on soils datasets downloaded from SLGA - mostly
about setting NA correctly
}
\keyword{internal}
