% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tile_aoi}
\alias{tile_aoi}
\title{Tile AOI}
\usage{
tile_aoi(aoi = NULL, size = 1)
}
\arguments{
\item{aoi}{sf bbox already converted from user input and aligned to requrested
product}

\item{size}{Number, max side length of tiles in decimal degrees. Default 1.}
}
\value{
a list of bounding boxes subdividing the area of interest. Return
list is not yet aligned to product (handled later).
}
\description{
If an AOI is Large, chop it up so that a series of GETs can be sent
}
\keyword{internal}
