% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{transform_bb}
\alias{transform_bb}
\title{transform a bounding box}
\usage{
transform_bb(bbox = NULL, crs = 4283)
}
\arguments{
\item{bbox}{Object of class 'bbox' generated by
\code{\link[sf:st_bbox]{sf::st_bbox}}.}

\item{crs}{EPSG code of desired coordinate reference system.}
}
\value{
'bbox' object with the desired CRS.
}
\description{
Transforms an `sf` style bounding box object into the specified coordinate
reference system.
}
\note{
This is a stopgap until and unless a dedicated sf method for bbox
transformations is written. Currently used to coerce to WGS84 under specific
circumstances. Since this helper was written for an Australia-specific package,
it is assumed that it will be used to transform coordinates from a fairly
limited list of known crs's, chiefly GDA94 (EPSG:4283) and its associated
UTM and Albers equal area projections.
}
\keyword{internal}
