% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validate_aoi}
\alias{validate_aoi}
\title{Validate AOI}
\usage{
validate_aoi(aoi = NULL, product = NULL)
}
\arguments{
\item{aoi}{length 4 numeric vector, `raster` object, or `sf` object.}

\item{product}{Character, one of the options from column 'Code' in
`slga_product_info`.}
}
\description{
Checks that an area of interest is of appropriate projection, size, and extent.
}
\keyword{Internal}
