% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{align_aoi}
\alias{align_aoi}
\title{align bounding box}
\usage{
align_aoi(aoi = NULL, product = NULL, snap = "out")
}
\arguments{
\item{aoi}{`sf` bbox object}

\item{product}{Character, one of the options from column 'Short_Name' in
\code{\link[slga:slga_product_info]{slga_product_info}}.}

\item{snap}{Character; 'near', 'in', or 'out'. Defaults to 'out'.}
}
\description{
Cheerfully stolen from code{\link[raster:alignExtent]{raster::alignExtent}}
and adapted to sf bbox objects. This is designed to prevent WCS server-side
interpolation.
}
\keyword{internal}
