% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_slga_data}
\alias{get_slga_data}
\title{GET soil or landscape data}
\usage{
get_slga_data(url = NULL, out_temp = NULL)
}
\arguments{
\item{url}{valid output from \code{\link[slga:make_soils_url]{make_soils_url}}
or \code{\link[slga:make_lscape_url]{make_lscape_url}}}

\item{out_temp}{location to write content - valid file path with .tif
extension.}
}
\value{
httr \code{\link[httr:response]{response()}} object with content
  stored on disk
}
\description{
Quietly sends a httr GET request to an SLGA web service endpoint.
}
\keyword{internal}
