% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slga.R
\docType{package}
\name{slga-package}
\alias{slga}
\alias{slga-package}
\title{slga: Data Access Tools for the Soil and Landscape Grid of Australia}
\description{
Provides access to Soil and Landscape Grid of Australia raster datasets via
existing Open Geospatial Consortium Web Coverage Services. See
\url{https://www.clw.csiro.au/aclep/soilandlandscapegrid/index.html}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/obrl-soil/slga}
  \item Report bugs at \url{https://github.com/obrl-soil/slga/issues}
}

}
\author{
\strong{Maintainer}: Lauren O'Brien \email{obrlsoilau@gmail.com} (\href{https://orcid.org/0000-0002-7336-2171}{ORCID})

Other contributors:
\itemize{
  \item Ross Searle \email{ross.searle@csiro.au} (\href{https://orcid.org/0000-0003-0256-1496}{ORCID}) [bibliographic antecedent]
}

}
