% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slga_attribute_info}
\alias{slga_attribute_info}
\title{SLGA Attribute Information}
\format{
A data frame with 14 observations and 4 variables \describe{
\item{Name}{Attribute name}
\item{Code}{Short code for attribute}
\item{Units}{Attribute measurement units}
\item{Transformation}{Attribute measurement scaling}
\item{NAT}{Whether the attribute is available as part of this product.}
\item{NAT_3D}{Whether the attribute is available as part of this product.}
\item{SA}{Whether the attribute is available as part of this product.}
\item{TAS}{Whether the attribute is available as part of this product.}
\item{WA}{Whether the attribute is available as part of this product.}
}
}
\source{
See also
\url{https://www.clw.csiro.au/aclep/soilandlandscapegrid/ProductDetails-SoilAttributes.html}
}
\usage{
slga_attribute_info
}
\description{
A data frame containing information about the modelled soils attributes
available from the Soil and Landscape Grid of Australia.
}
\keyword{datasets}
