% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_processing.R
\name{tidy_lscape_data}
\alias{tidy_lscape_data}
\title{tidy landscape parameter rasters}
\usage{
tidy_lscape_data(r = NULL, product = NULL, write_out = FALSE, filedir)
}
\arguments{
\item{r}{raster object downloaded over WCS}

\item{product}{character; the SLGA terrain product in question}

\item{write_out}{Boolean, whether to write the retrieved dataset to disk.
Defaults to FALSE.}

\item{filedir}{directory in which to write files if write_out == TRUE.}
}
\value{
a raster, but a better one
}
\description{
Does some post-processing on landscape datasets downloaded from SLGA - mostly
about setting NA correctly
}
\keyword{internal}
