% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validate_aoi}
\alias{validate_aoi}
\title{Validate AOI}
\usage{
validate_aoi(aoi = NULL, product = NULL)
}
\arguments{
\item{aoi}{Numeric; bounding coordinates or an `sf` or `raster` object from
which they can be derived.}

\item{product}{Character, one of the options from column 'Short_Name' in
\code{\link[slga:slga_product_info]{slga_product_info}}.}
}
\description{
Checks that an area of interest is of appropriate projection, size, and
extent.
}
\keyword{internal}
